@@ -5,8 +5,8 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "proxy": "ng serve --proxy proxy/config.json --host 0.0.0.0",
+    "start": "ng serve --aot",
+    "proxy": "ng serve --aot --proxy proxy/config.json --host 0.0.0.0",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
@@ -14,24 +14,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/platform-server": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "bootstrap-sass": "^3.3.7",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "lz-string": "^1.4.4",
     "moment": "^2.18.1",
-    "ngx-bootstrap": "^1.7.1",
-    "rxjs": "^5.0.1",
+    "ngx-bootstrap": "^2.0.0-beta.7",
+    "rxjs": "^5.5.2",
     "sugar": "^2.0.4",
     "sugar-language": "^2.0.4",
     "ts-helpers": "^1.1.1",
@@ -39,7 +38,7 @@
   },
   "devDependencies": {
     "@angular/cli": "^1.0.0",
-    "@angular/compiler": "^4.1.3",
+    "@angular/compiler": "^5.0.0",
     "@types/jasmine": "2.5.38",
     "@types/lz-string": "^1.3.32",
     "@types/node": "^6.0.42",
@@ -54,6 +53,6 @@
     "protractor": "~4.0.13",
     "ts-node": "1.2.1",
     "tslint": "^4.3.0",
-    "typescript": "2.3.4"
+    "typescript": "2.4.2"
   }
 }
