@@ -1,8 +1,9 @@
 import { Observable } from "rxjs";
+import { filter, map, tap } from 'rxjs/operators';
 import { Injectable, Inject, InjectionToken } from "@angular/core";
-import { Response } from "@angular/http";
 import { Router, NavigationStart } from "@angular/router";
 import { RequestHook } from "calico/core/api.service";
+import { HttpResponse } from "@angular/common/http";
 
 export const VERSION_INFO = new InjectionToken<VersionInfo>("versionInfo");
 
@@ -18,14 +19,16 @@ export class VersionCheckHook implements RequestHook {
     private router: Router
   ) {
   }
-  apply(url: string, form: any, observable: Observable<Response>): Observable<Response> {
-      return observable.do(res => {
-        let version = res.headers.get(this.versionInfo.key);
-        if (version == null || version == this.versionInfo.currentVersion) return;
-        this.router.events
-          .filter(e => e instanceof NavigationStart)
-          .map(e => e as NavigationStart)
-          .subscribe(loc => location.href = loc.url);
-      });
+  apply(url: string, form: any, observable: Observable<HttpResponse<any>>): Observable<HttpResponse<any>> {
+    return tap((res: HttpResponse<any>) => {
+      let version = res.headers.get(this.versionInfo.key);
+      if (version == null || version == this.versionInfo.currentVersion) return;
+      let events = filter.call(this.router.events, (e: any) => e instanceof NavigationStart);
+      console.log(events);
+      events = map.call(events, (e: any) => e as NavigationStart);
+      console.log(events);
+      events.subscribe((loc: any) => location.href = loc.url);
+      console.log(events);
+    })(observable);
   }
 }
